# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HUTCH Resolve planned API",
    "version": "1.0.0",
    "description": "Shared Resolve wire contract v1.0.0. Implemented routes include health/readiness, session lifecycle, customer account/case/investigation APIs, action/receipt, agent review, and signed Voice session/callback endpoints. Conversation turns depend on Tevin's injected in-process controller; full frontends and live Voice qualification remain pending. See docs/contracts.md and context.md for status."
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local Resolve runtime"
    }
  ],
  "paths": {
...
    "/cases/{id}": {
      "get": {
        "operationId": "get_case",
        "summary": "get case",
        "description": "Proposed interface; not implemented at documentation baseline.",
        "tags": [
          "Resolve"
        ],
        "x-owner": "Harry",
        "security": [
          {
            "CustomerSession": []
          },
          {
            "AgentSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseView"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Correlation ID"
              }
            }
          },
          "401": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/ResolveRealm"
          }
        ]
      }
    },
    "/cases/{id}/investigations": {
      "post": {
        "operationId": "investigate_case",
        "summary": "investigate case",
        "description": "Proposed interface; not implemented at documentation baseline.",
        "tags": [
          "Resolve"
        ],
        "x-owner": "Harry",
        "security": [
          {
            "CustomerSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationResult"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Correlation ID"
              }
            }
          },
          "401": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationRequest"
              }
            }
          }
        }
      }
    },
...
    "/cases/{id}/receipt": {
      "get": {
        "operationId": "get_receipt",
        "summary": "get receipt",
        "description": "Proposed interface; not implemented at documentation baseline.",
        "tags": [
          "Resolve"
        ],
        "x-owner": "Harry",
        "security": [
          {
            "CustomerSession": []
          },
          {
            "AgentSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptView"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Correlation ID"
              }
            }
          },
          "401": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "revision",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "$ref": "#/components/parameters/ResolveRealm"
          }
        ]
      }
    },
...
    "/agent/cases": {
      "get": {
        "operationId": "list_agent_cases",
        "summary": "list agent cases",
        "description": "Review queue: REVIEW_REQUIRED case, accepted escalation, or existing agent review history. Run-scoped. Sort updated_at DESC then ID DESC.",
        "tags": [
          "Agent"
        ],
        "x-owner": "Harry",
        "security": [
          {
            "AgentSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseQueue"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Correlation ID"
              }
            }
          },
          "401": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "review_status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReviewStatus"
            }
          },
          {
            "name": "complaint_type",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ComplaintType"
            }
          },
          {
            "name": "evidence_state",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EvidenceState"
            }
          },
          {
            "name": "delivery_state",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/DeliveryState"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          }
        ]
      }
    },
    "/agent/cases/{case_id}": {
      "get": {
        "operationId": "get_agent_case",
        "summary": "get agent case",
        "description": "Proposed interface; not implemented at documentation baseline.",
        "tags": [
          "Agent"
        ],
        "x-owner": "Harry",
        "security": [
          {
            "AgentSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCaseDetail"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Correlation ID"
              }
            }
          },
          "401": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Error; see shared contract codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CustomerSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "resolve_customer_session"
      },
      "AgentSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "resolve_agent_session"
      },
      "VoiceSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Voice-Signature",
        "description": "HMAC SHA256 of timestamp.event_id.body_sha256; all companion headers mandatory; 60-second skew."
      }
    },
    "schemas": {
...
      "ComplaintType": {
        "type": "string",
        "enum": [
          "BALANCE_RECHARGE",
          "DATA_DEPLETION",
          "CONNECTIVITY",
          "VAS_DISPUTE"
        ]
      },
      "EvidenceState": {
        "type": "string",
        "enum": [
          "SUFFICIENT",
          "PARTIAL",
          "CONFLICTING"
        ]
      },
...
      "ReviewStatus": {
        "type": "string",
        "enum": [
          "NEW",
          "IN_REVIEW",
          "CLOSED"
        ]
      },
      "DeliveryState": {
        "type": "string",
        "enum": [
          "PENDING",
          "DELIVERED",
          "FAILED",
          "REVIEW_REQUIRED"
        ]
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "request_id": {
                "type": "string",
                "format": "uuid"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            },
            "required": [
              "code",
              "message",
              "retryable",
              "request_id",
              "details"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Finding": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "evidence_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "code",
          "text",
          "evidence_ids"
        ]
      },
...
      "InvestigationRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1
          },
          "complaint_type": {
            "$ref": "#/components/schemas/ComplaintType"
          },
          "window_start": {
            "type": "string",
            "format": "date-time"
          },
          "window_end": {
            "type": "string",
            "format": "date-time"
          },
          "reported_facts": {
            "$ref": "#/components/schemas/ReportedFacts"
          }
        },
        "required": [
          "expected_version",
          "complaint_type",
          "window_start",
          "window_end",
          "reported_facts"
        ]
      },
      "InvestigationResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "case_id": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "integer",
            "minimum": 1
          },
          "complaint_type": {
            "$ref": "#/components/schemas/ComplaintType"
          },
          "window_start": {
            "type": "string",
            "format": "date-time"
          },
          "window_end": {
            "type": "string",
            "format": "date-time"
          },
          "evidence_state": {
            "$ref": "#/components/schemas/EvidenceState"
          },
          "findings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Finding"
            }
          },
          "calculations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Calculation"
            }
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            }
          },
          "source_status": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceStatus"
            }
          },
          "missing": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "conflicts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "eligible_actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EligibleAction"
            }
          },
          "review_reasons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "simulation": {
            "type": "boolean",
            "const": true
          }
        },
        "required": [
          "id",
          "case_id",
          "revision",
          "complaint_type",
          "window_start",
          "window_end",
          "evidence_state",
          "findings",
          "calculations",
          "evidence",
          "source_status",
          "missing",
          "conflicts",
          "eligible_actions",
          "review_reasons",
          "created_at",
          "simulation"
        ]
      },
...
      "ReceiptReference": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "id",
          "revision"
        ]
      },
      "CaseView": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "conversation_id": {
            "type": "string",
            "format": "uuid"
          },
          "account_id": {
            "type": "string",
            "format": "uuid"
          },
          "complaint_type": {
            "$ref": "#/components/schemas/ComplaintType"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "AWAITING_CUSTOMER",
              "ACTION_PENDING",
              "REVIEW_REQUIRED",
              "RESOLVED"
            ]
          },
          "review_status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "investigation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvestigationResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "operation_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "receipt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReceiptReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "simulation": {
            "type": "boolean",
            "const": true
          }
        },
        "required": [
          "id",
          "conversation_id",
          "account_id",
          "complaint_type",
          "status",
          "review_status",
          "version",
          "created_at",
          "updated_at",
          "investigation",
          "operation_ids",
          "receipt",
          "simulation"
        ]
      },
...
      "CaseQueue": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CaseQueueRow"
            }
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "items",
          "next_cursor"
        ]
      },
...
    },
    "parameters": {
      "ResolveRealm": {
        "name": "X-Resolve-Realm",
        "in": "header",
        "required": false,
        "description": "Selects the browser session realm for shared customer/agent reads. Required when both cookies are present.",
        "schema": {
          "type": "string",
          "enum": [
            "customer",
            "agent"
          ]
        }
      }
    }
  }
}